refactor lsk and lsk base class
