+++ .editorconfig
@@ -8,8 +8,8 @@ charset = utf-8
 end_of_line = lf
 insert_final_newline = true
 trim_trailing_whitespace = true
-indent_style = tab
-tab_width = 4
+indent_style = space
+tab_width = 2
 [*.md]
 max_line_length = off
 trim_trailing_whitespace = false
